220801 : A308M v1.01 (成) 1. 修改A308讀取機制     - Client:         - 在接收Server狀態回應後，發出NS_SEND_INFO_ACK訊號讓Server端知道Client已收到回應         - 修改讀取機制，未即時回應讀取設備狀態指令的Server如在讀取其他裝置的流程中有回應狀態，會將該Server加入此次讀取流程中     - Server         - 在發送完設備訊息後，若未接收到Client發出的NS_SEND_INFO_ACK訊號，每2秒會再發送一次設備訊息給Client，避免訊號丟失         - 從休眠狀態喚醒後，多等待2秒再讀取A308數值(設定義A308_SLEEP_MODE=1)。 2. A308來源表減少讀取內容，以減少傳送時間 3. 增加NODE_INFO Property，用以統一回傳設備狀態(for APP，未測試) 4. 增加可在一般狀態下存取設定參數的指令:NODE_SENSOR_SETUP_GET/NODE_SENSOR_SETUP_SET

+++ user/bus_drv/bus_usart.h
@@ -117,9 +117,9 @@
 #define TIME_OUT_COUNT_RX        WAIT_MS(20)    //20ms
 
 
-extern uchar volatile CounterRx,CounterTx;
+extern uint16 volatile CounterRx,CounterTx;
 extern uchar *RxBuff;//[];
-extern uchar UsartCounterTx,UsartCounterRx;
+extern uint16 UsartCounterTx,UsartCounterRx;
 
 void UsartInit(void);
 void UsartDeInit(void);
